\name{postcode-package}
\alias{postcode-package}
\alias{postcode}
\docType{package}
\title{Post code database of mainland China}
\description{
This package contains a database of Chinese post code, 
which includes the post code, province, city and district.
}
\details{
\tabular{ll}{
Package: \tab postcode\cr
Maintainer: \tab Guangyang Li <bot@guangyangli.com>\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-12\cr
License: \tab GPL-2\cr
}
}
\author{
Guangyang Li
}
\keyword{ package }